src/api/algorithm/creation/requirements.txt
# public
# diffusers
# pip install git+https://github.com/huggingface/diffusers.git -i https://pypi.org/simple
torch
transformers
accelerate
sentencepiece
protobuf
fastapi
uvicorn
loguru
python-dotenv
pytz
# audio
soundfile
torchsde
# video
opencv-python
imageio
imageio-ffmpeg
# tts
# ChatTTS
chattts
# # CosyVoice-300M-SFT
# hyperpyyaml
# onnxruntime-gpu
# # onnxruntime
# torchaudio